+++ sefaria/model/text.py
@@ -29,6 +29,7 @@ from .schema import deserialize_tree, SchemaNode, VirtualNode, DictionaryNode, J
 from sefaria.system.database import db
 
 import sefaria.system.cache as scache
+from sefaria.system.cache import InMemoryCache
 from sefaria.system.exceptions import InputError, BookNameError, PartialRefInputError, IndexSchemaError, \
     NoVersionFoundError, DictionaryEntryNotFoundError
 from sefaria.utils.hebrew import is_hebrew, hebrew_term
@@ -4613,6 +4614,8 @@ class Library(object):
         self._title_regex_strings = {}
         self._title_regexes = {}
         Ref.clear_cache()
+        in_memory_cache = InMemoryCache()
+        in_memory_cache.reset_all()
         if include_toc:
             self.rebuild_toc()
