+++ sefaria/system/cache.py
@@ -131,8 +131,24 @@ def delete_template_cache(fragment_name='', *args):
     delete_cache_elem('template.cache.%s.%s' % (fragment_name, hashlib.md5(':'.join([arg for arg in args]).encode('utf-8')).hexdigest()))
 
 
-class InMemoryCache:
-    def reset(self, key, value):
-        set_shared_cache_elem(key, value)
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls._instances.get(cls) is None:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+class InMemoryCache(metaclass=Singleton):
+    data = {}
+    def set(self, key, val):
+        self.data[key] = val
+
+    def get(self, key):
+        return self.data.get(key,  None)
+
+    def reset_all(self):
+        for k in self.data:
+            self.data[k] = None
